remove deprecated split view controller delegate methods
in iOS 8, the primary view controller ([svc.viewControllers firstObject]) is
always hanging around. it is set to the delegate so that it can decide how
to handle the difference between a collapsed and expanded experience. at
this stage, the secondary view controller is always discarded. this isn't
what the intended functionality is, rather if an item is selected then the
collapsed view should show the secondary view controller collapsed into the
primary.